# public/swagger-docs.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "description": "Quizapp API",
            "url": "https://quizzapp.sbsa.uz"
        }
    ],
    "info": {
        "version": "1.0.0",
        "title": "QuizApp Back-end API",
        "description": "API for managing users, cards, and card items in the QuizApp application.",
        "contact": {
            "name": "Muhammadali Yuldoshev",
            "url": "http://yuldoshev.uz",
            "email": "mukhammadaliweb@gmail.com"
        }
    },
    "paths": {
...
        "/api/category/{slug}": {
            "parameters": [
                {
                    "schema": {
                        "type": "string"
                    },
                    "name": "slug",
                    "in": "path",
                    "required": true
                }
            ],
            "get": {
                "summary": "Get Category By Slug",
                "tags": [
                    "Category"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "statusCode": {
                                                    "type": "integer"
                                                },
                                                "timestamp": {
                                                    "type": "string"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "category": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "string"
                                                        },
                                                        "title": {
                                                            "type": "string"
                                                        },
                                                        "slug": {
                                                            "type": "string"
                                                        },
                                                        "parentId": {
                                                            "type": "string"
                                                        },
                                                        "createdAt": {
                                                            "type": "string"
                                                        },
                                                        "updatedAt": {
                                                            "type": "string"
                                                        },
                                                        "children": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {}
                                                            }
                                                        },
                                                        "cards": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string"
                                                                    },
                                                                    "authorId": {
                                                                        "type": "string"
                                                                    },
                                                                    "title": {
                                                                        "type": "string"
                                                                    },
                                                                    "slug": {
                                                                        "type": "string"
                                                                    },
                                                                    "categoryId": {
                                                                        "type": "string"
                                                                    },
                                                                    "createdAt": {
                                                                        "type": "string"
                                                                    },
                                                                    "updatedAt": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    },
                                    "x-examples": {
                                        "Example 1": {
                                            "ok": true,
                                            "meta": {
                                                "statusCode": 200,
                                                "timestamp": "2024-09-12T10:16:29.069Z",
                                                "path": "/api/category/ozbek-tili-1726135110767",
                                                "message": "Request successful"
                                            },
                                            "data": {
                                                "category": {
                                                    "id": "3da37bf9-9932-4314-973d-1313da4577e6",
                                                    "title": "O'zbek tili",
                                                    "slug": "ozbek-tili-1726135110767",
                                                    "parentId": null,
                                                    "createdAt": "2024-09-12T09:58:30.767Z",
                                                    "updatedAt": "2024-09-12T09:58:30.767Z",
                                                    "children": [],
                                                    "cards": [
                                                        {
                                                            "id": "df18bcd0-4594-409e-b439-095cffdf4572",
                                                            "authorId": "8587605f-6d74-4272-ac1a-784d6d829bb0",
                                                            "title": "O'zbekcha yangi lug'atlar to'plami",
                                                            "slug": "ozbekcha-yangi-lugatlar-toplami-1726119642211",
                                                            "categoryId": "3da37bf9-9932-4314-973d-1313da4577e6",
                                                            "createdAt": "2024-09-12T05:40:42.212Z",
                                                            "updatedAt": "2024-09-12T10:16:08.229Z"
                                                        }
                                                    ]
                                                }
                                            }
                                        }
                                    }
                                },
                                "examples": {
                                    "Example": {
                                        "value": {
                                            "ok": true,
                                            "meta": {
                                                "statusCode": 200,
                                                "timestamp": "2024-09-12T10:16:29.069Z",
                                                "path": "/api/category/ozbek-tili-1726135110767",
                                                "message": "Request successful"
                                            },
                                            "data": {
                                                "category": {
                                                    "id": "3da37bf9-9932-4314-973d-1313da4577e6",
                                                    "title": "O'zbek tili",
                                                    "slug": "ozbek-tili-1726135110767",
                                                    "parentId": null,
                                                    "createdAt": "2024-09-12T09:58:30.767Z",
                                                    "updatedAt": "2024-09-12T09:58:30.767Z",
                                                    "children": [],
                                                    "cards": [
                                                        {
                                                            "id": "df18bcd0-4594-409e-b439-095cffdf4572",
                                                            "authorId": "8587605f-6d74-4272-ac1a-784d6d829bb0",
                                                            "title": "O'zbekcha yangi lug'atlar to'plami",
                                                            "slug": "ozbekcha-yangi-lugatlar-toplami-1726119642211",
                                                            "categoryId": "3da37bf9-9932-4314-973d-1313da4577e6",
                                                            "createdAt": "2024-09-12T05:40:42.212Z",
                                                            "updatedAt": "2024-09-12T10:16:08.229Z"
                                                        }
                                                    ]
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                },
                                "examples": {
                                    "Example": {
                                        "value": {
                                            "ok": false,
                                            "meta": {
                                                "statusCode": 401,
                                                "timestamp": "2024-09-12T10:12:53.427Z",
                                                "path": "/api/category/ozbek-tili-1725965808397",
                                                "errors": []
                                            },
                                            "message": "Unauthorized"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "statusCode": {
                                                    "type": "integer"
                                                },
                                                "timestamp": {
                                                    "type": "string"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "errors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {}
                                                    }
                                                }
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "x-examples": {
                                        "Example 1": {
                                            "ok": false,
                                            "meta": {
                                                "statusCode": 404,
                                                "timestamp": "2024-09-12T10:11:43.916Z",
                                                "path": "/api/category/ozbek-tili-1725965808397",
                                                "errors": []
                                            },
                                            "message": "Category not found"
                                        }
                                    }
                                },
                                "examples": {
                                    "Example": {
                                        "value": {
                                            "ok": false,
                                            "meta": {
                                                "statusCode": 404,
                                                "timestamp": "2024-09-12T10:11:43.916Z",
                                                "path": "/api/category/ozbek-tili-1725965808397",
                                                "errors": []
                                            },
                                            "message": "Category not found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "get-api-category-slug"
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {},
        {
            "Authorization": []
        }
    ]
}